FEAT: modularize the `knowFor` field in actors
