+++ services/actors/search/types.ts
@@ -1,3 +1,5 @@
+import { KnownFor } from "services/actors/types/knowFor";
+
 export interface ActorSearchResponse {
   page: number;
   results: Result[];
@@ -14,36 +16,4 @@ export interface Result {
   name: string;
   popularity: number;
   profile_path: null | string;
-}
-
-export interface KnownFor {
-  backdrop_path?: null | string;
-  first_air_date?: string;
-  genre_ids: number[];
-  id: number;
-  media_type: MediaType;
-  name?: string;
-  origin_country?: string[];
-  original_language: OriginalLanguage;
-  original_name?: string;
-  overview: string;
-  poster_path: null | string;
-  vote_average: number;
-  vote_count: number;
-  adult?: boolean;
-  original_title?: string;
-  release_date?: string;
-  title?: string;
-  video?: boolean;
-}
-
-export enum MediaType {
-  Movie = "movie",
-  Tv = "tv",
-}
-
-export enum OriginalLanguage {
-  En = "en",
-  Ko = "ko",
-  Zh = "zh",
 }
